Extract the abort-marker subsystem into a composed AbortMarkerSubsystem class
Third incremental fileSyncConnection.ts extraction. The authenticated abort marker's state (tokens, write inputs, the decision one-shot and its grace timer) and logic (the memoized bounded write, the seal, the fail-closed authenticated verify) move to packages/core/src/connection/abortMarker.ts as a class the connection composes. Collaborators arrive by constructor injection (log, live role accessor, the transport-budget primitives, which stay owned by the connection); per-call data arrives as parameters, so the subsystem holds no reference to the connection. External behavior is byte-identical and the connection keeps thin delegating members (armAbort, writeAbortMarker, sealAbort, abortArmed) for its CLI and test callers. The leftover-marker sweep stays in the poll loop where it belongs.
